Simulate Normal Population Scores for Multiple Items

This function allows you to create normal populations for data that would include repeated measures items. Additionally, the data can be rounded and/or truncated to ensure it matches a target scale - for example, a 1-7 type rating scale.

Usage

calculate_proportion(samples, cutoff, grouping_items = NULL, score)

Arguments

samples

The bootstrapped samples from your population

cutoff

The cutoff score for an item to be well measured from the standard errors of your items

grouping_items

The names of columns to group your data by for the cutoff calculation, usually this column is the item column

score

The column of the score you wish to calculate for your cutoff score SE

Value

summary_DF

Summary of the proportion of items below the standard error cutoff score.

Examples

# step 1 create data like what I think I'll get or use your own
pops <- simulate_population(mu = 4, mu_sigma = .2, sigma = 2,
                            sigma_sigma = .2, number_items = 30, number_scores = 20,
                            smallest_sigma = .02, min_score = 1, max_score = 7, digits = 0)

# step 2 calculate our cut off score
cutoff <- calculate_cutoff(population = pops,
 grouping_items = "item",
 score = "score",
 minimum = 1,
 maximum = 7)

cutoff$se_items
#>  [1] 0.3732856 0.3741657 0.4507304 0.3086473 0.3469794 0.3718729 0.4827280
#>  [8] 0.4672146 0.5155477 0.4222434 0.4103913 0.3692097 0.4259664 0.3524351
#> [15] 0.4784844 0.4723959 0.3933660 0.4000000 0.4284304 0.3830281 0.3940345
#> [22] 0.3026201 0.3589092 0.4328607 0.3900067 0.3879772 0.4095890 0.3857665
#> [29] 0.4546543 0.4184873
cutoff$sd_items
#> [1] 0.05014582
cutoff$cutoff
#>       40% 
#> 0.3891949 
cutoff$prop_var
#> [1] 0.01671527

# step 3 simulate bootstrapped samples
samples <- bootstrap_samples(start = 20, stop = 100,
 increase = 5, population = pops,
 replace = TRUE, grouping_items = "item")

# step 4 and 5
proportion_summary <- calculate_proportion(samples = samples,
 cutoff = cutoff$cutoff,
 grouping_items = "item",
 score = "score")
proportion_summary
#> # A tibble: 17 × 2
#>    sample_size percent_below
#>          <dbl>         <dbl>
#>  1          20         0.489
#>  2          25         0.734
#>  3          30         0.905
#>  4          35         0.978
#>  5          40         0.998
#>  6          45         1    
#>  7          50         1    
#>  8          55         1    
#>  9          60         1    
#> 10          65         1    
#> 11          70         1    
#> 12          75         1    
#> 13          80         1    
#> 14          85         1    
#> 15          90         1    
#> 16          95         1    
#> 17         100         1