Import Datasets from the Linguistic Annotated Bibliography

This function allows you to import the current datasets available from Semantic Priming GitHub Group. These datasets generally have psycholinguistic variables like Age of Acquisition, length, neighborhoods, Familiarity, Concretenes, and more. We will update datasets as we have them. All data belongs to the person who published them.

Usage

import_lab(bibtexID = NULL, citation = NULL, language = NULL, variables = NULL)

Arguments

bibtexID

The bibtex ID of the dataset you are trying to load. You can leave all parameters blank to load just the metadata.

citation

Include the citation for the dataset you loaded - will only load if you include a bibtex ID, use `TRUE` to get the citation.

language

If you include a bibtex ID, you will get back the language of the dataset, if you do not include a bibtex ID, it will return a list of datasets in that language.

variables

If you include a bibtex ID, you will get back the variables included the dataset, if you do not include a bibtex ID, it will return a list of datasets that include that variable (can also be paired with language). Use the column names from the metadata as your filter.

Value

metadataThe metadata list of avaliable datasets

loaded_dataThe dataset you requested to load

languageThe language of the dataset you requested to load

variablesThe variables of the dataset you requested to load

datasetsPossible datasets based on your language and variable names

Examples

# import_lab()
# import_lab(bibtexID = "Birchenough2017", citation = TRUE)
# import_lab(language = "English", variables = c("aoa", "freq"))