Import the SPAML Data or Stimuli

This function allows you to import the current datasets available from the SPAML project. You can use `data("primeData")` to view the list of available datasets/materials.

Usage

import_prime(filename)

Arguments

filename

Include the filename of the dataset you want to load. Note: some of the `full_data` datasets are very large and require a good bit of memory to load.

Value

a dataset or stimuli list in a dataframe.

Examples

# ar_words <- import_prime("ar_words.csv")