SPAML Priming Data Download

Dataset to use the import_prime function to import data or materials from the SPAML project.

Usage

data(primeData)

Format

A data frame of links and information about the SPAML project.

type

which type of dataset you would like to view/load.

filename

the name of the file of the dataset.

language

the language of the stimuli/materials in the data.

location

a download link for the most recent data.

Please note: we are currently on v1.0.1 for the data from the project.