subs2vec Project Model and Frequency Data Downloads

Dataset to use the import_subs function to import subtitle fastText model outputs and frequency counts. Includes information about matching `udpipe` models for tagging.

Usage

data(subsData)

Format

A data frame of links and information about the subs2vec project.

language_codethe two letter language code of the model

subs_veca link to download the subtitle only fastText model

subs_counta link to download the frequencies for the tokens in the subtitle data

wiki_veca link to download the wikipedia only fastText model

wiki_counta link to download the frequencies for the tokens in the wikipedia data

filesthe number of files in the OpenSubtitles data

tokensthe number of tokens in the OpenSubtitles data

sentencesthe number of sentences in the OpenSubtitles data

languagethe full name of the language for reference

udpipe_modelthe matching `udpipe` model for download to parse tokens